+++ src/main/java/pl/asie/computronics/network/NetworkHandlerClient.java
@@ -7,7 +7,8 @@ import net.minecraft.tileentity.TileEntity;
 import pl.asie.computronics.Computronics;
 import pl.asie.computronics.api.audio.AudioPacketClientHandler;
 import pl.asie.computronics.api.audio.AudioPacketRegistry;
-import pl.asie.computronics.item.ItemPortableTapeDrive;
+import pl.asie.computronics.item.ItemPortableTapeDrive.PortableDriveManager;
+import pl.asie.computronics.item.ItemPortableTapeDrive.TapeDrive;
 import pl.asie.computronics.oc.driver.DriverCardNoise;
 import pl.asie.computronics.oc.driver.DriverCardSoundBase;
 import pl.asie.computronics.reference.Mods;
@@ -90,7 +91,15 @@ public class NetworkHandlerClient extends MessageHandlerBase {
 			}
 			break;
 			case PORTABLE_TAPE_STOP: {
-				ItemPortableTapeDrive.PortableDriveManager.INSTANCE.stopTapeDrive(packet.readString());
+				PortableDriveManager.INSTANCE.stopTapeDrive(packet.readString());
+			}
+			break;
+			case PORTABLE_TAPE_STATE: {
+				TapeDrive tapeDrive = PortableDriveManager.INSTANCE.getTapeDrive(packet.readString());
+				State state = State.VALUES[packet.readUnsignedByte()];
+				if(tapeDrive != null) {
+					tapeDrive.switchState(state);
+				}
 			}
 			break;
 		}